chore: refactor links
